+++ CHANGELOG.md
@@ -32,6 +32,11 @@
 
 - Fixed an issue where Snowpark pandas plugin would unconditionally disable `AutoSwitchBackend` even when users had explicitly configured it via environment variables or programmatically.
 
+### Snowpark pandas API Updates
+
+#### Improvements
+- Add a new configuration variable to control transfer costs out of Snowflake when using hybrid execution. Lower the default to 100k from 10M.
+
 ## 1.35.0 (2025-07-24)
 
 ### Snowpark Python API Updates